requirements.txt
line-bot-sdk
flask~=2.1.2

requests~=2.27.1
SQLAlchemy~=1.4.37
python-dotenv~=0.20.0